This is the initial version of the rule-base AI software Proposition.
The purpose of this software is to generate data from the input data
(one database table from another one) based on arbitrary user-editable
rules. Currently it is capable of creating a .csv file from another .csv
file. It runs in command-line only. Intended for Windows 7 and newer but
can be compiled for Linux distros such as Debian and Ubuntu.
There are two main ideas for further development:
1) Adding a graphical user interface.
2) Adding actual database access, such as the ability to read data from
a view or table in MySQL or MSSQL and write the output into another
database table.